feat: add basic API setup using axum and aide for docs
To pass the source and target instances to the REST handler functions
using the `.with_state` on the axum ApiRouter, we had to change the
containers for most of the objects in the interfaces from rc::Rc (and
some Boxes) to std::sync::Arc, because we now need to have the Sync and
Send trait implemented for the source and target. Otherwise, AppState
wouldn't be Clonable and thus couldn't be used for the axum state.