refactor(source/keycloak): store attributes instead of lib struct
When we introduced the WebhookUser / WebhookGroup we had to construct a
UserRepresentation / GroupRepresentation object from the Keycloak
library. This was tedious, as we had to specify many fields with None.
This also made it rather unsafe as it introduces the potential of using
fields which a normal Keycloak representation struct would have set but
which wasn't set via the webhook representation.

Hence, we decided to not store the library representation anymore and
instead store selected attributes which we need for the application
logic manually. This allows us to convert the Webhook object into the
internal objects properly.